Add sync to ffi boundary
